add latest buy endpoints

+++ Application/Service/ProductService/Dto/ProductUpdateDto.cs
@@ -10,7 +10,7 @@ namespace JWT_NET_5.Application.Service.ProductService.Dto
 		[Required]
 		public int Amount { get; set; }
 		[Required]
-		public int Cost { get; set; }
+		public Double Cost { get; set; }
 		[Required]
 		public Guid UserId { get; set; }
 	}
